@@ -1,16 +1,35 @@
 module github.com/deepshore/docker-machine-driver-proxmoxve
 
-go 1.16
+go 1.20
 
 require (
-	github.com/docker/docker v20.10.24+incompatible // indirect
-	github.com/docker/machine v0.16.2
-	github.com/labstack/gommon v0.3.1
-	github.com/mattn/go-colorable v0.1.12 // indirect
-	github.com/moby/term v0.0.0-20210619224110-3f7ff695adc6 // indirect
-	github.com/mosolovsa/go_cat_sshfilerw v0.0.0-20170307095915-3f6912dbd9e4
-	golang.org/x/crypto v0.0.0-20220525230936-793ad666bf5e
-	golang.org/x/net v0.0.0-20220526153639-5463443f8c37 // indirect
-	golang.org/x/term v0.0.0-20220526004731-065cf7ba2467 // indirect
-	gopkg.in/resty.v1 v1.12.0
+	github.com/labstack/gommon v0.4.1
+	github.com/luthermonson/go-proxmox v0.0.0-beta2
+	github.com/rancher/machine v0.15.0-rancher98
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
+
+require (
+	github.com/Azure/go-ansiterm v0.0.0-20210617225240-d185dfc1b5a1 // indirect
+	github.com/buger/goterm v1.0.4 // indirect
+	github.com/diskfs/go-diskfs v1.2.0 // indirect
+	github.com/docker/docker v20.10.2+incompatible // indirect
+	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/jinzhu/copier v0.3.4 // indirect
+	github.com/magefile/mage v1.14.0 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/moby/term v0.5.0 // indirect
+	github.com/stretchr/testify v1.8.4
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasttemplate v1.2.2 // indirect
+	golang.org/x/crypto v0.0.0-20220411220226-7b82a4e95df4 // indirect
+	golang.org/x/sys v0.14.0 // indirect
+	golang.org/x/term v0.0.0-20210927222741-03fcf44c2211 // indirect
+	gopkg.in/djherbis/times.v1 v1.2.0 // indirect
 )
