@@ -1,44 +1,35 @@
 package main
 
 import (
-	"bytes"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"encoding/pem"
+	"context"
+	"crypto/tls"
 	"errors"
 	"fmt"
-	"io/ioutil"
-	mrand "math/rand"
+	"math/rand"
+	"net/http"
 	"net/url"
 	"os"
-	"regexp"
 	"strconv"
 	"strings"
 	"time"
 
-	"golang.org/x/crypto/ssh"
-	resty "gopkg.in/resty.v1"
-
-	sshrw "github.com/mosolovsa/go_cat_sshfilerw"
-
-	"github.com/docker/machine/libmachine/drivers"
-	"github.com/docker/machine/libmachine/mcnflag"
-	mssh "github.com/docker/machine/libmachine/ssh"
-	"github.com/docker/machine/libmachine/state"
 	"github.com/labstack/gommon/log"
+	"github.com/luthermonson/go-proxmox"
+
+	"github.com/rancher/machine/libmachine/drivers"
+	"github.com/rancher/machine/libmachine/mcnflag"
+	"github.com/rancher/machine/libmachine/ssh"
+	"github.com/rancher/machine/libmachine/state"
 )
 
 // Driver for Proxmox VE
 type Driver struct {
 	*drivers.BaseDriver
-	driver *ProxmoxVE
-
-	// Top-level strategy for proisioning a new node
-	ProvisionStrategy string
+	client *proxmox.Client
 
 	// Basic Authentication for Proxmox VE
 	Host     string // Host to connect to
+	Port     string // Port to connect to (default 8006)
 	Node     string // optional, node to create VM on, host used if omitted but must match internal node name
 	User     string // username
 	Password string // password
@@ -58,8 +49,6 @@ type Driver struct {
 	Citype          string // Specifies the cloud-init configuration format.
 	NUMA            string // Enable/disable NUMA
 
-	CiEnabled string
-
 	NetModel    string // Net Interface Model, [e1000, virtio, realtek, etc...]
 	NetFirewall string // Enable/disable firewall
 	NetMtu      string // set nic MTU
@@ -69,9 +58,8 @@ type Driver struct {
 	ScsiController string
 	ScsiAttributes string
 
-	VMID          string // VM ID only filled by create()
+	VMID          int    // VM ID only filled by create()
 	VMIDRange     string // acceptable range of VMIDs
-	VMUUID        string // UUID to confirm
 	CloneVMID     string // VM ID to clone
 	CloneFull     int    // Make a full (detached) clone from parent (defaults to true if VMID is not a template, otherwise false)
 	GuestUsername string // user to log into the guest OS to copy the public key
@@ -81,7 +69,18 @@ type Driver struct {
 	CPUSockets    string // The number of cpu sockets.
 	CPUCores      string // The number of cores per socket.
 	driverDebug   bool   // driver debugging
-	restyDebug    bool   // enable resty debugging
+}
+
+// NewDriver returns a new driver
+func NewDriver(hostName, storePath string) drivers.Driver {
+	return &Driver{
+		BaseDriver: &drivers.BaseDriver{
+			SSHUser:     "docker",
+			MachineName: hostName,
+			StorePath:   storePath,
+		},
+		Citype: "nocloud", // default to nocloud since this driver will only support linux
+	}
 }
 
 func (d *Driver) debugf(format string, v ...interface{}) {
@@ -96,22 +95,37 @@ func (d *Driver) debug(v ...interface{}) {
 	}
 }
 
-func (d *Driver) connectAPI() error {
-	if d.driver == nil {
-		d.debugf("Create called")
+func (d *Driver) connectApi() (client *proxmox.Client, err error) {
+	var options []proxmox.Option
 
-		d.debugf("Connecting to %s as %s@%s with password '%s'", d.Host, d.User, d.Realm, d.Password)
-		c, err := GetProxmoxVEConnectionByValues(d.User, d.Password, d.Realm, d.Host)
-		d.driver = c
-		if err != nil {
-			return fmt.Errorf("Could not connect to host '%s' with '%s@%s'", d.Host, d.User, d.Realm)
-		}
-		if d.restyDebug {
-			c.EnableDebugging()
-		}
-		d.debugf("Connected to PVE version '" + d.driver.Version + "'")
+	options = append(options, proxmox.WithHTTPClient(&http.Client{
+		Transport: &http.Transport{
+			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+		},
+	}))
+	credentials := proxmox.Credentials{
+		Username: d.User,
+		Password: d.Password,
+		Realm:    d.Realm,
 	}
-	return nil
+	options = append(options, proxmox.WithCredentials(&credentials))
+
+	proxmoxUrl := fmt.Sprintf("https://%s:%s/api2/json", d.Host, d.Port)
+	log.Debug(fmt.Sprintf("Connecting to %s", proxmoxUrl))
+	d.client = proxmox.NewClient(proxmoxUrl, options...)
+
+	version, err := d.client.Version(context.Background())
+	if err != nil {
+		return nil, err
+	}
+	c, err2 := d.client.Cluster(context.Background())
+	if err2 != nil {
+		return nil, err2
+	}
+
+	log.Infof("Connected to pve cluster %s with version: %s", c.Name, version.Version)
+
+	return d.client, err
 }
 
 // GetCreateFlags returns the argument flags for the program
@@ -124,16 +138,16 @@ func (d *Driver) GetCreateFlags() []mcnflag.Flag {
 			Value:  "192.168.1.253",
 		},
 		mcnflag.StringFlag{
+			EnvVar: "PROXMOXVE_PROXMOX_PORT",
+			Name:   "proxmoxve-proxmox-port",
+			Usage:  "Port to connect to",
+			Value:  "8006",
+		},
+		mcnflag.StringFlag{
 			EnvVar: "PROXMOXVE_PROXMOX_NODE",
 			Name:   "proxmoxve-proxmox-node",
 			Usage:  "Node to use (defaults to host)",
 			Value:  "",
-		},
-		mcnflag.StringFlag{
-			EnvVar: "PROXMOXVE_PROVISION_STRATEGY",
-			Name:   "proxmoxve-provision-strategy",
-			Usage:  "Provision strategy (cdrom|clone)",
-			Value:  "cdrom",
 		},
 		mcnflag.StringFlag{
 			EnvVar: "PROXMOXVE_PROXMOX_USER_NAME",
@@ -226,16 +240,10 @@ func (d *Driver) GetCreateFlags() []mcnflag.Flag {
 			Value:  "",
 		},
 		mcnflag.StringFlag{
-			EnvVar: "PROXMOXVE_VM_CLONE_VNID",
+			EnvVar: "PROXMOXVE_VM_CLONE_VMID",
 			Name:   "proxmoxve-vm-clone-vmid",
 			Usage:  "vmid to clone",
 			Value:  "",
-		},
-		mcnflag.IntFlag{
-			EnvVar: "PROXMOXVE_VM_CLONE_FULL",
-			Name:   "proxmoxve-vm-clone-full",
-			Usage:  "make a full clone or not (0=false, 1=true, 2=use proxmox default logic",
-			Value:  2,
 		},
 		mcnflag.StringFlag{
 			EnvVar: "PROXMOXVE_VM_START_ONBOOT",
@@ -247,18 +255,6 @@ func (d *Driver) GetCreateFlags() []mcnflag.Flag {
 			EnvVar: "PROXMOXVE_VM_PROTECTION",
 			Name:   "proxmoxve-vm-protection",
 			Usage:  "protect the VM and disks from removal (0=false, 1=true, ''=default)",
-			Value:  "", // leave the flag default value blank to support the clone default behavior if not explicity set of 'use what is most appropriate'
-		},
-		mcnflag.StringFlag{
-			EnvVar: "PROXMOXVE_VM_CITYPE",
-			Name:   "proxmoxve-vm-citype",
-			Usage:  "cloud-init type (nocloud|configdrive2)",
-			Value:  "", // leave the flag default value blank to support the clone default behavior if not explicity set of 'use what is most appropriate'
-		},
-		mcnflag.StringFlag{
-			EnvVar: "PROXMOXVE_VM_CIENABLED",
-			Name:   "proxmoxve-vm-cienabled",
-			Usage:  "cloud-init enabled (implied with clone strategy 0=false, 1=true, ''=default)",
 			Value:  "", // leave the flag default value blank to support the clone default behavior if not explicity set of 'use what is most appropriate'
 		},
 		mcnflag.StringFlag{
@@ -316,32 +312,11 @@ func (d *Driver) GetCreateFlags() []mcnflag.Flag {
 			Value:  22,
 		},
 		mcnflag.BoolFlag{
-			EnvVar: "PROXMOXVE_DEBUG_RESTY",
-			Name:   "proxmoxve-debug-resty",
-			Usage:  "enables the resty debugging",
-		},
-		mcnflag.BoolFlag{
 			EnvVar: "PROXMOXVE_DEBUG_DRIVER",
 			Name:   "proxmoxve-debug-driver",
 			Usage:  "enables debugging in the driver",
 		},
 	}
-}
-
-func (d *Driver) ping() bool {
-	if d.driver == nil {
-		return false
-	}
-
-	command := NodesNodeQemuVMIDAgentPostParameter{Command: "ping"}
-	err := d.driver.NodesNodeQemuVMIDAgentPost(d.Node, d.VMID, &command)
-
-	if err != nil {
-		d.debug(err)
-		return false
-	}
-
-	return true
 }
 
 // DriverName returns the name of the driver
@@ -353,10 +328,12 @@ func (d *Driver) DriverName() string {
 func (d *Driver) SetConfigFromFlags(flags drivers.DriverOptions) error {
 	d.debug("SetConfigFromFlags called")
 
-	d.ProvisionStrategy = flags.String("proxmoxve-provision-strategy")
-
 	// PROXMOX API Connection settings
 	d.Host = flags.String("proxmoxve-proxmox-host")
+	d.Port = flags.String("proxmoxve-proxmox-port")
+	if len(d.Port) == 0 {
+		d.Port = "8006"
+	}
 	d.Node = flags.String("proxmoxve-proxmox-node")
 	if len(d.Node) == 0 {
 		d.Node = d.Host
@@ -374,11 +351,8 @@ func (d *Driver) SetConfigFromFlags(flags drivers.DriverOptions) error {
 	d.Memory *= 1024
 	d.VMIDRange = flags.String("proxmoxve-vm-vmid-range")
 	d.CloneVMID = flags.String("proxmoxve-vm-clone-vmid")
-	d.CloneFull = flags.Int("proxmoxve-vm-clone-full")
 	d.Onboot = flags.String("proxmoxve-vm-start-onboot")
 	d.Protection = flags.String("proxmoxve-vm-protection")
-	d.Citype = flags.String("proxmoxve-vm-citype")
-	d.CiEnabled = flags.String("proxmoxve-vm-cienabled")
 	d.ImageFile = flags.String("proxmoxve-vm-image-file")
 	d.CPUSockets = flags.String("proxmoxve-vm-cpu-sockets")
 	d.CPU = flags.String("proxmoxve-vm-cpu")
@@ -390,25 +364,12 @@ func (d *Driver) SetConfigFromFlags(flags drivers.DriverOptions) error {
 	d.NetVlanTag = flags.Int("proxmoxve-vm-net-tag")
 	d.ScsiController = flags.String("proxmoxve-vm-scsi-controller")
 	d.ScsiAttributes = flags.String("proxmoxve-vm-scsi-attributes")
+	d.driverDebug = flags.Bool("proxmoxve-debug-driver")
 
 	//SSH connection settings
 	d.GuestSSHPort = flags.Int("proxmoxve-ssh-port")
 	d.GuestUsername = flags.String("proxmoxve-ssh-username")
 	d.GuestPassword = flags.String("proxmoxve-ssh-password")
-
-	//SWARM Settings
-	d.SwarmMaster = flags.Bool("swarm-master")
-	d.SwarmHost = flags.String("swarm-host")
-
-	//Debug option
-	d.driverDebug = flags.Bool("proxmoxve-debug-driver")
-	d.restyDebug = flags.Bool("proxmoxve-debug-resty")
-
-	if d.restyDebug {
-		d.debug("enabling Resty debugging")
-		resty.SetLogger(log.Output())
-		resty.SetDebug(true)
-	}
 
 	return nil
 }
@@ -440,23 +401,150 @@ func (d *Driver) GetNetVlanTag() int {
 	return d.NetVlanTag
 }
 
+func (d *Driver) GetNode(nodeName string) (*proxmox.Node, error) {
+	if d.client == nil {
+		client, err := d.connectApi()
+		if err != nil {
+			return nil, err
+		}
+		d.client = client
+	}
+
+	n, err := d.client.Node(context.Background(), nodeName)
+	if err != nil {
+		return nil, err
+	}
+	return n, err
+}
+
+func (d *Driver) ConfigureVM(name string, value string) error {
+	d.debugf("ConfigureVM: %s %s", name, value)
+	vm, err := d.GetVM()
+	if err != nil {
+		return err
+	}
+	var config proxmox.VirtualMachineOption
+	config.Name = name
+	config.Value = value
+
+	configTask, err2 := vm.Config(context.Background(), config)
+
+	if err2 != nil {
+		return err2
+	}
+
+	// wait for the config task
+	if err4 := configTask.Wait(context.Background(), time.Duration(5*time.Second), time.Duration(300*time.Second)); err4 != nil {
+		return err4
+	}
+
+	d.debugf("Config task finished")
+
+	return nil
+}
+
+func (d *Driver) OperateVM(operation string) error {
+
+	vm, err := d.GetVM()
+	if err != nil {
+		return err
+	}
+
+	switch operation {
+	case "start":
+		task, err2 := vm.Start(context.Background())
+		log.Debug(task.ID)
+		if err2 != nil {
+			return err2
+		}
+		// wait for the task
+		if err3 := task.Wait(context.Background(), time.Duration(5*time.Second), time.Duration(300*time.Second)); err3 != nil {
+			return err3
+		}
+	case "stop":
+		task, err2 := vm.Stop(context.Background())
+		log.Debug(task.ID)
+		if err2 != nil {
+			return err2
+		}
+		// wait for the task
+		if err3 := task.Wait(context.Background(), time.Duration(5*time.Second), time.Duration(300*time.Second)); err3 != nil {
+			return err3
+		}
+	case "kill":
+		task, err2 := vm.Stop(context.Background())
+		log.Debug(task.ID)
+		if err2 != nil {
+			return err2
+		}
+		// wait for the task
+		if err3 := task.Wait(context.Background(), time.Duration(5*time.Second), time.Duration(300*time.Second)); err3 != nil {
+			return err3
+		}
+	case "restart":
+		task, err2 := vm.Reset(context.Background())
+		log.Debug(task.ID)
+		if err2 != nil {
+			return err2
+		}
+		// wait for the task
+		if err3 := task.Wait(context.Background(), time.Duration(5*time.Second), time.Duration(300*time.Second)); err3 != nil {
+			return err3
+		}
+	default:
+		return errors.New("Invalid operation: " + operation)
+	}
+
+	return err
+
+}
+
+func (d *Driver) GetVM() (*proxmox.VirtualMachine, error) {
+	d.debugf("GetVM issued")
+	if d.VMID < 1 {
+		return nil, errors.New("invalid VMID")
+	}
+
+	n, err := d.GetNode(d.Node)
+	d.debugf("GetNode returned: %v", n)
+	if err != nil {
+		return nil, err
+	}
+	vm, err2 := n.VirtualMachine(context.Background(), d.VMID)
+	if err2 != nil {
+		return nil, err2
+	}
+	d.debugf("GetVM returned: %v", vm)
+	return vm, err
+}
+
 // GetIP returns the ip
 func (d *Driver) GetIP() (string, error) {
-	d.connectAPI()
+	vm, err := d.GetVM()
 
-	ip, err := d.driver.GetEth0IPv4(d.Node, d.VMID)
-	if err != nil {
-		// TODO: should we return the cached IP here?
-		return ip, err
+	if err := vm.WaitForAgent(context.Background(), 300); err != nil {
+		return "", err
+	}
+	net := vm.VirtualMachineConfig.Net0
+	iFaces, err3 := vm.AgentGetNetworkIFaces(context.Background())
+	if err3 != nil {
+		return "", err3
+	}
+	for _, iface := range iFaces {
+		if strings.Contains(strings.ToLower(net), strings.ToLower(iface.HardwareAddress)) {
+			for _, ip := range iface.IPAddresses {
+				if ip.IPAddressType == "ipv4" {
+					d.IPAddress = ip.IPAddress
+				}
+			}
+		}
 	}
 
-	// set/update IP if success
-	if d.IPAddress != ip {
-		d.debugf("driver IP is set as '%s'", ip)
-		d.IPAddress = ip
+	if d.IPAddress == "" {
+		return "", err
 	}
 
-	return ip, err
+	return d.IPAddress, err
 }
 
 // GetSSHHostname returns the ssh host returned by the API
@@ -476,127 +564,35 @@ func (d *Driver) GetSSHUsername() string {
 
 // GetState returns the state of the VM
 func (d *Driver) GetState() (state.State, error) {
-	if len(d.VMID) < 1 {
-		return state.Error, errors.New("invalid VMID")
-	}
-
-	err := d.connectAPI()
+	vm, err := d.GetVM()
 	if err != nil {
-		return state.Error, err
+		return state.None, err
 	}
 
-	// sanity check the UUID
-	config, err := d.driver.GetConfig(d.Node, d.VMID)
-	if err != nil {
-		return state.Error, err
+	if err := vm.Ping(context.Background()); err != nil {
+		return state.None, err
 	}
 
-	cVMMUUID := getUUIDFromSmbios1(config.Data.Smbios1)
-	if len(d.VMUUID) > 1 && d.VMUUID != cVMMUUID {
-		return state.Error, fmt.Errorf("UUID mismatch - %s (stored) vs %s (current)", d.VMUUID, cVMMUUID)
+	if vm.IsStopped() {
+		return state.Stopped, nil
 	}
 
-	//Starting
-	//Running
-	//Stopped
-	//Paused
-	//Error
-
-	ip, err := d.GetIP()
-
-	// only consider fully running when IP is available
-	if err == nil && len(ip) > 0 {
+	if vm.IsRunning() {
 		return state.Running, nil
 	}
 
-	// starting if qemu-guest-agent is available but no IP
-	if d.ping() {
-		return state.Starting, nil
-	}
-
-	pveState, err := d.driver.NodesNodeQemuVMIDStatusCurrentGet(d.Node, d.VMID)
-	if err != nil {
-		return state.Error, err
-	}
-
-	switch pveState {
-	case "stopped":
-		return state.Stopped, nil
-	case "running":
-		return state.Starting, nil
-	}
-
-	return state.Error, fmt.Errorf("unkown error detecting VM state")
+	return state.None, nil
 }
 
 // PreCreateCheck is called to enforce pre-creation steps
 func (d *Driver) PreCreateCheck() error {
 
-	err := d.connectAPI()
-	if err != nil {
-		return err
-	}
-
-	switch d.ProvisionStrategy {
-	case "cdrom":
-		// set defaults for cdrom
-		// replicating pre-clone behavior of setting a default on the parameter
-		if len(d.Storage) < 1 {
-			d.Storage = "local"
-		}
-
-		if len(d.StorageType) < 1 {
-			d.StorageType = "raw"
-		}
-
-		if len(d.NetBridge) < 1 {
-			d.NetBridge = "vmbr0"
-		}
-
-		if len(d.GuestUsername) < 1 {
-			d.GuestUsername = "docker"
-		}
-
-		if len(d.GuestPassword) < 1 {
-			d.GuestPassword = "tcuser"
-		}
-
-		// prepare StorageFilename
-		switch d.StorageType {
-		case "raw":
-			fallthrough
-		case "qcow2":
-			break
-		default:
-			return fmt.Errorf("storage type '%s' is not supported", d.StorageType)
-		}
-
-		storageType, err := d.driver.GetStorageType(d.Node, d.Storage)
+	if d.client == nil {
+		client, err := d.connectApi()
 		if err != nil {
 			return err
 		}
-
-		filename := "-disk-0"
-		switch storageType {
-		case "lvmthin":
-			fallthrough
-		case "zfs":
-			fallthrough
-		case "ceph":
-			if d.StorageType != "raw" {
-				return fmt.Errorf("type '%s' on storage '%s' does only support raw", storageType, d.Storage)
-			}
-		case "nfs":
-			fallthrough
-		case "dir":
-			filename += "." + d.StorageType
-		}
-		// this is not the finale filename, it'll be constructed in Create()
-		d.StorageFilename = filename
-	case "clone":
-		break
-	default:
-		return fmt.Errorf("invalid provision strategy '%s'", d.ProvisionStrategy)
+		d.client = client
 	}
 
 	return nil
@@ -605,289 +601,102 @@ func (d *Driver) PreCreateCheck() error {
 // Create creates a new VM with storage
 func (d *Driver) Create() error {
 
-	// create and save a new SSH key pair
-	d.debug("creating new ssh keypair")
-	key, err := d.createSSHKey()
+	newId, err6 := d.GetVmidInRange()
+	if err6 != nil {
+		return err6
+	}
+
+	clone := &proxmox.VirtualMachineCloneOptions{
+		Name:    d.MachineName,
+		Full:    1,
+		Pool:    d.Pool,
+		Format:  d.StorageType,
+		Storage: d.Storage,
+		NewID:   newId,
+	}
+
+	d.debugf("cloning new vm from template id '%s'", d.CloneVMID)
+
+	node, err := d.client.Node(context.Background(), d.Node)
 	if err != nil {
 		return err
 	}
 
-	// !! Workaround for MC-7982.
-	key = strings.TrimSpace(key)
-	key = fmt.Sprintf("%s %s-%d", key, d.MachineName, time.Now().Unix())
-	// !! End workaround for MC-7982.
-
-	// add some random wait time here to help with race conditions in the proxmox api
-	// the app could be getting invoked several times in rapid succession so some small waits may be helpful
-	mrand.Seed(time.Now().UnixNano()) // Seed the random number generator using the current time (nanoseconds since epoch)
-	r := mrand.Intn(5000)
-	d.debugf("sleeping %d milliseconds before retrieving next ID", r)
-	time.Sleep(time.Duration(r) * time.Millisecond)
-
-	// get next available VMID
-	// NOTE: we want to lock in the ID as quickly as possible after retrieving (ie: invoke QemuPost or Clone ASAP to avoid race conditions with other instances)
-	d.debug("Retrieving next ID")
-
-	var id string
-	if len(d.VMIDRange) > 0 {
-		VMIDParts := strings.Split(d.VMIDRange, ":")
-		low, err := strconv.Atoi(VMIDParts[0])
-		if err != nil {
-			return err
-		}
-
-		var high = 0
-		if len(VMIDParts) > 1 {
-			high, err = strconv.Atoi(VMIDParts[1])
-			if err != nil {
-				return err
-			}
-		}
-
-		d.debugf("Looking for available VMID in range '%d:%d'", low, high)
-
-		var i = low
-		for len(id) < 1 {
-			id, err = d.driver.ClusterNextIDGet(i)
-			if err != nil && high > 0 && i > high {
-				return err
-			}
-
-			if high > 0 && i >= high {
-				return fmt.Errorf("no VMIDs available in range '%d:%d'", low, high)
-			}
-
-			i++
-		}
-	} else {
-		id, err = d.driver.ClusterNextIDGet(0)
-		if err != nil {
-			return err
-		}
-	}
-
-	d.debugf("Next ID is '%s'", id)
-	d.VMID = id
-
-	switch d.ProvisionStrategy {
-	case "cdrom":
-
-		// prefixing StorageFilename with VMID
-		d.StorageFilename = "vm-" + d.VMID + d.StorageFilename
-
-		volume := NodesNodeStorageStorageContentPostParameter{
-			Filename: d.StorageFilename,
-			Size:     d.DiskSize + "G",
-			VMID:     d.VMID,
-		}
-
-		d.debugf("Creating disk volume '%s' with size '%s'", volume.Filename, volume.Size)
-		diskname, err := d.driver.NodesNodeStorageStorageContentPost(d.Node, d.Storage, &volume)
-		if err != nil {
-			return err
-		}
-
-		if !strings.HasSuffix(diskname, d.StorageFilename) {
-			return fmt.Errorf("returned diskname is not correct: should be '%s' but was '%s'", d.StorageFilename, diskname)
-		}
-
-		npp := NodesNodeQemuPostParameter{
-			VMID:       d.VMID,
-			Agent:      "1",
-			Autostart:  "1",
-			Onboot:     d.Onboot,
-			Memory:     d.Memory,
-			Sockets:    d.CPUSockets,
-			Cores:      d.CPUCores,
-			SCSI0:      d.StorageFilename,
-			Ostype:     "l26",
-			Name:       d.BaseDriver.MachineName,
-			KVM:        "1", // if you test in a nested environment, you may have to change this to 0 if you do not have nested virtualization
-			Scsihw:     d.ScsiController,
-			Cdrom:      d.ImageFile,
-			Pool:       d.Pool,
-			Protection: d.Protection,
-		}
-
-		if d.CiEnabled == "1" {
-			npp.Citype = d.Citype
-			npp.Ide3 = d.Storage + ":cloudinit"
-		}
-
-		if len(d.ScsiAttributes) > 0 {
-			npp.SCSI0 += "," + d.ScsiAttributes
-		}
-
-		if len(d.NUMA) > 0 {
-			npp.NUMA = d.NUMA
-		}
-
-		if len(d.CPU) > 0 {
-			npp.CPU = d.CPU
-		}
-
-		npp.Net0, _ = d.generateNetString()
-
-		if d.StorageType == "qcow2" {
-			npp.SCSI0 = d.Storage + ":" + d.VMID + "/" + volume.Filename
-		} else if d.StorageType == "raw" {
-			if strings.HasSuffix(volume.Filename, ".raw") {
-				// raw files (having .raw) should have the VMID in the path
-				npp.SCSI0 = d.Storage + ":" + d.VMID + "/" + volume.Filename
-			} else {
-				npp.SCSI0 = d.Storage + ":" + volume.Filename
-			}
-		}
-		d.debugf("Creating VM '%s' with '%d' of memory", npp.VMID, npp.Memory)
-		taskid, err := d.driver.NodesNodeQemuPost(d.Node, &npp)
-		if err != nil {
-			return err
-		}
-
-		err = d.driver.WaitForTaskToComplete(d.Node, taskid)
-		if err != nil {
-			return err
-		}
-
-		if d.CiEnabled == "1" {
-			// specially handle setting sshkeys
-			// https://forum.proxmox.com/threads/how-to-use-pvesh-set-vms-sshkeys.52570/
-			taskid, err = d.driver.NodesNodeQemuVMIDConfigSetSSHKeys(d.Node, d.VMID, key)
-			if err != nil {
-				return err
-			}
-
-			err = d.driver.WaitForTaskToComplete(d.Node, taskid)
-			if err != nil {
-				return err
-			}
-		}
-
-		break
-	case "clone":
-
-		// clone
-		clone := NodesNodeQemuVMIDClonePostParameter{
-			Newid: d.VMID,
-			Name:  d.BaseDriver.MachineName,
-			Pool:  d.Pool,
-		}
-
-		switch d.CloneFull {
-		case 0:
-			clone.Full = "0"
-			break
-		case 1:
-			clone.Full = "1"
-			clone.Format = d.StorageType
-			clone.Storage = d.Storage
-			break
-		case 2:
-			clone.Format = d.StorageType
-			clone.Storage = d.Storage
-			break
-		}
-
-		d.debugf("cloning template id '%s' as vmid '%s'", d.CloneVMID, clone.Newid)
-
-		taskid, err := d.driver.NodesNodeQemuVMIDClonePost(d.Node, d.CloneVMID, &clone)
-		if err != nil {
-			return err
-		}
-
-		err = d.driver.WaitForTaskToComplete(d.Node, taskid)
-		if err != nil {
-			return err
-		}
-
-		// resize
-		resize := NodesNodeQemuVMIDResizePutParameter{
-			Disk: "scsi0",
-			Size: d.DiskSize + "G",
-		}
-		d.debugf("resizing disk '%s' on vmid '%s' to '%s'", resize.Disk, d.VMID, resize.Size)
-
-		err = d.driver.NodesNodeQemuVMIDResizePut(d.Node, d.VMID, &resize)
-		if err != nil {
-			return err
-		}
-
-		// set config values
-		d.debugf("setting VM config values for vmid '%s'", d.VMID)
-		npp := NodesNodeQemuPostParameter{
-			Agent:      "1",
-			Autostart:  "1",
-			Memory:     d.Memory,
-			Sockets:    d.CPUSockets,
-			Cores:      d.CPUCores,
-			KVM:        "1", // if you test in a nested environment, you may have to change this to 0 if you do not have nested virtualization,
-			Citype:     d.Citype,
-			Onboot:     d.Onboot,
-			Protection: d.Protection,
-		}
-
-		if len(d.NetBridge) > 0 {
-			npp.Net0, _ = d.generateNetString()
-		}
-
-		if len(d.NUMA) > 0 {
-			npp.NUMA = d.NUMA
-		}
-
-		if len(d.CPU) > 0 {
-			npp.CPU = d.CPU
-		}
-
-		taskid, err = d.driver.NodesNodeQemuVMIDConfigPost(d.Node, d.VMID, &npp)
-		if err != nil {
-			return err
-		}
-
-		// append newly minted ssh key to existing (if any)
-		d.debugf("retrieving existing cloud-init sshkeys from vmid '%s'", d.VMID)
-		config, err := d.driver.GetConfig(d.Node, d.CloneVMID)
-		if err != nil {
-			return err
-		}
-
-		var SSHKeys string
-
-		if len(config.Data.SSHKeys) > 0 {
-			SSHKeys, err = url.QueryUnescape(config.Data.SSHKeys)
-			if err != nil {
-				return err
-			}
-
-			SSHKeys = strings.TrimSpace(SSHKeys)
-			SSHKeys += "\n"
-		}
-
-		SSHKeys += key
-		SSHKeys = strings.TrimSpace(SSHKeys)
-
-		// specially handle setting sshkeys
-		// https://forum.proxmox.com/threads/how-to-use-pvesh-set-vms-sshkeys.52570/
-		taskid, err = d.driver.NodesNodeQemuVMIDConfigSetSSHKeys(d.Node, d.VMID, SSHKeys)
-		if err != nil {
-			return err
-		}
-
-		err = d.driver.WaitForTaskToComplete(d.Node, taskid)
-		if err != nil {
-			return err
-		}
-		break
-	default:
-		return fmt.Errorf("invalid provision strategy '%s'", d.ProvisionStrategy)
-	}
-
-	// Set the newly minted UUID
-	config, err := d.driver.GetConfig(d.Node, d.VMID)
+	cloneVmId, err := strconv.Atoi(d.CloneVMID)
 	if err != nil {
 		return err
 	}
-	d.VMUUID = getUUIDFromSmbios1(config.Data.Smbios1)
-	d.debugf("VM created with uuid '%s'", d.VMUUID)
+
+	clonevm, err := node.VirtualMachine(context.Background(), cloneVmId)
+	if err != nil {
+		return err
+	}
+
+	_, task, err := clonevm.Clone(context.Background(), clone)
+	d.debugf("clone task for new vmid '%d' created", newId)
+
+	if err != nil {
+		return err
+	}
+
+	// wait for the clone task
+	if err := task.Wait(context.Background(), time.Duration(5*time.Second), time.Duration(300*time.Second)); err != nil {
+		return err
+	}
+	d.debugf("clone finished for vmid '%d'", newId)
+
+	// explicity set vmid after clone completion to be sure
+	d.VMID = newId
+
+	d.debugf("vmid values VMID: '%d'", d.VMID)
+
+	// resize
+	d.debugf("resizing disk '%s' on vmid '%s' to '%s'", "scsi0", d.VMID, d.DiskSize+"G")
+
+	vm, err4 := d.GetVM()
+	if err4 != nil {
+		return err4
+	}
+	err5 := vm.ResizeDisk(context.Background(), "scsi0", d.DiskSize+"G")
+	if err5 != nil {
+		return err5
+	}
+
+	d.debugf("add misc configuration options")
+
+	d.ConfigureVM("agent", "1")
+	d.ConfigureVM("autostart", "1")
+	d.ConfigureVM("memory", fmt.Sprint(d.Memory))
+	d.ConfigureVM("sockets", d.CPUSockets)
+	d.ConfigureVM("cores", d.CPUCores)
+	d.ConfigureVM("kvm", "1")
+	d.ConfigureVM("citype", d.Citype)
+	d.ConfigureVM("onboot", d.Onboot)
+	d.ConfigureVM("protection", d.Protection)
+
+	if len(d.NetBridge) > 0 {
+		d.ConfigureVM("net0", d.generateNetString())
+	}
+
+	if len(d.NUMA) > 0 {
+		d.ConfigureVM("numa", d.NUMA)
+	}
+
+	if len(d.CPU) > 0 {
+		d.ConfigureVM("cpu", d.CPU)
+	}
+
+	// append newly minted ssh key to existing (if any)
+	SSHKeys, err2 := d.appendVmSshKeys(vm)
+	if err2 != nil {
+		return err2
+	}
+
+	err3 := d.ConfigureVM("sshkeys", SSHKeys)
+	d.debugf("cloud-init sshkeys set to '%s'", SSHKeys)
+	if err3 != nil {
+		return err3
+	}
 
 	// start the VM
 	err = d.Start()
@@ -895,78 +704,59 @@ func (d *Driver) Create() error {
 		return err
 	}
 
-	// let VM start a settle a little
-	d.debugf("waiting for VM to start, wait 10 seconds")
-	time.Sleep(10 * time.Second)
-
-	// wait for qemu-guest-agent
-	err = d.waitForQemuGuestAgent()
+	// wait for the agent and get the IPAddress
+	vmIp, err := d.GetIP()
 	if err != nil {
 		return err
 	}
 
-	// wait for network to come up
-	err = d.waitForNetwork()
-
-	// set the IPAddress
-	_, err = d.GetIP()
-	if err != nil {
-		return err
-
-	}
-
-	switch d.ProvisionStrategy {
-	case "cdrom":
-		if len(d.GuestPassword) > 0 {
-			return d.prepareSSHWithPassword()
-		}
-		return nil
-	case "clone":
-		fallthrough
-	default:
-		return nil
-	}
-}
-
-func (d *Driver) waitForQemuGuestAgent() error {
-	d.debugf("waiting for VM qemu-guest-agent to start")
-	d.connectAPI()
-	for !d.ping() {
-		d.debugf("waiting for VM qemu-guest-agent to start")
-		time.Sleep(5 * time.Second)
-	}
+	d.debugf("VM got an IP: %s", vmIp)
 
 	return nil
 }
 
-func (d *Driver) waitForNetwork() error {
-	d.debugf("waiting for VM network to start")
-	d.connectAPI()
+func (d *Driver) appendVmSshKeys(vm *proxmox.VirtualMachine) (string, error) {
+	// create and save a new SSH key pair
+	d.debug("creating new ssh keypair")
+	key, err := d.createSSHKey()
+	if err != nil {
+		return "", err
+	}
 
-	var up = false
-	var ip string
-	var err error
+	// !! Workaround for MC-7982.
+	key = strings.TrimSpace(key)
+	key = fmt.Sprintf("%s %s-%d", key, d.MachineName, time.Now().Unix())
+	// !! End workaround for MC-7982.
 
-	for !up {
-		ip, err = d.driver.GetEth0IPv4(d.Node, d.VMID)
+	var SSHKeys string
+
+	if len(vm.VirtualMachineConfig.SSHKeys) > 0 {
+		SSHKeys, err = url.QueryUnescape(vm.VirtualMachineConfig.SSHKeys)
 		if err != nil {
-			d.debugf("waiting for VM network to start")
-			time.Sleep(5 * time.Second)
-		} else {
-			if len(ip) > 0 {
-				up = true
-				d.debugf("VM network started with ip: %s", ip)
-			} else {
-				d.debugf("waiting for VM network to start")
-				time.Sleep(5 * time.Second)
-			}
+			return "", err
 		}
+
+		SSHKeys = strings.TrimSpace(SSHKeys)
+		SSHKeys += "\n"
 	}
 
-	return nil
+	SSHKeys += key
+	SSHKeys = strings.TrimSpace(SSHKeys)
+
+	// specially handle setting sshkeys
+	// https://forum.proxmox.com/threads/how-to-use-pvesh-set-vms-sshkeys.52570/
+
+	d.debugf("retrieving existing cloud-init sshkeys from vmid '%s'", d.VMID)
+
+	r := strings.NewReplacer("+", "%2B", "=", "%3D", "@", "%40")
+
+	SSHKeys = url.PathEscape(SSHKeys)
+	SSHKeys = r.Replace(SSHKeys)
+
+	return SSHKeys, nil
 }
 
-func (d *Driver) generateNetString() (string, error) {
+func (d *Driver) generateNetString() string {
 	var net string = fmt.Sprintf("model=%s,bridge=%s", d.NetModel, d.NetBridge)
 	if d.NetVlanTag != 0 {
 		net = fmt.Sprintf(net+",tag=%d", d.NetVlanTag)
@@ -980,315 +770,94 @@ func (d *Driver) generateNetString() (string, error) {
 		net = fmt.Sprintf(net+",mtu=%s", d.NetMtu)
 	}
 
-	return net, nil
-}
-
-func (d *Driver) prepareSSHWithPassword() error {
-	sshConfig := &ssh.ClientConfig{
-		User: d.GetSSHUsername(),
-		Auth: []ssh.AuthMethod{
-			ssh.Password(d.GuestPassword)},
-		HostKeyCallback: ssh.InsecureIgnoreHostKey(),
-	}
-
-	sshbasedir := "/home/" + d.GetSSHUsername() + "/.ssh"
-	hostname, _ := d.GetSSHHostname()
-	port, _ := d.GetSSHPort()
-	clientstr := fmt.Sprintf("%s:%d", hostname, port)
-
-	d.debugf("Creating directory '%s'", sshbasedir)
-	conn, err := ssh.Dial("tcp", clientstr, sshConfig)
-	if err != nil {
-		return err
-	}
-	session, err := conn.NewSession()
-	if err != nil {
-		return err
-	}
-
-	var stdoutBuf bytes.Buffer
-	session.Stdout = &stdoutBuf
-	session.Run("mkdir -p " + sshbasedir)
-	d.debugf(fmt.Sprintf("%s -> %s", hostname, stdoutBuf.String()))
-	session.Close()
-
-	d.debugf("Trying to copy to %s:%s", clientstr, sshbasedir)
-	c, err := sshrw.NewSSHclt(clientstr, sshConfig)
-	if err != nil {
-		return err
-	}
-
-	// Open a file
-	f, err := os.Open(d.GetSSHKeyPath() + ".pub")
-	if err != nil {
-		return err
-	}
-
-	// TODO: always fails with return status 127, but file was copied correclty
-	c.WriteFile(f, sshbasedir+"/authorized_keys")
-	// if err = c.WriteFile(f, sshbasedir+"/authorized_keys"); err != nil {
-	// 	d.debugf("Error on file write: ", err)
-	// }
-
-	// Close the file after it has been copied
-	defer f.Close()
-
-	return err
+	return net
 }
 
 // Start starts the VM
 func (d *Driver) Start() error {
-	if len(d.VMID) < 1 {
-		return errors.New("invalid VMID")
-	}
-
-	err := d.connectAPI()
-	if err != nil {
-		return err
-	}
-
-	// sanity check the UUID
-	config, err := d.driver.GetConfig(d.Node, d.VMID)
-	if err != nil {
-		return err
-	}
-
-	cVMMUUID := getUUIDFromSmbios1(config.Data.Smbios1)
-	if len(d.VMUUID) > 1 && d.VMUUID != cVMMUUID {
-		return fmt.Errorf("UUID mismatch - %s (stored) vs %s (current)", d.VMUUID, cVMMUUID)
-	}
-
-	taskid, err := d.driver.NodesNodeQemuVMIDStatusStartPost(d.Node, d.VMID)
-	if err != nil {
-		return err
-	}
-
-	err = d.driver.WaitForTaskToComplete(d.Node, taskid)
-
-	return err
+	return d.OperateVM("start")
 }
 
 // Stop stopps the VM
 func (d *Driver) Stop() error {
-	if len(d.VMID) < 1 {
-		return errors.New("invalid VMID")
-	}
-
-	err := d.connectAPI()
-	if err != nil {
-		return err
-	}
-
-	// sanity check the UUID
-	config, err := d.driver.GetConfig(d.Node, d.VMID)
-	if err != nil {
-		return err
-	}
-
-	cVMMUUID := getUUIDFromSmbios1(config.Data.Smbios1)
-	if len(d.VMUUID) > 1 && d.VMUUID != cVMMUUID {
-		return fmt.Errorf("UUID mismatch - %s (stored) vs %s (current)", d.VMUUID, cVMMUUID)
-	}
-
-	// shutdown
-	taskid, err := d.driver.NodesNodeQemuVMIDStatusShutdownPost(d.Node, d.VMID)
-	if err != nil {
-		return err
-	}
-
-	err = d.driver.WaitForTaskToComplete(d.Node, taskid)
-
-	return err
+	return d.OperateVM("stop")
 }
 
 // Restart restarts the VM
 func (d *Driver) Restart() error {
-	if len(d.VMID) < 1 {
-		return errors.New("invalid VMID")
-	}
-
-	err := d.connectAPI()
-	if err != nil {
-		return err
-	}
-
-	// sanity check the UUID
-	config, err := d.driver.GetConfig(d.Node, d.VMID)
-	if err != nil {
-		return err
-	}
-
-	cVMMUUID := getUUIDFromSmbios1(config.Data.Smbios1)
-	if len(d.VMUUID) > 1 && d.VMUUID != cVMMUUID {
-		return fmt.Errorf("UUID mismatch - %s (stored) vs %s (current)", d.VMUUID, cVMMUUID)
-	}
-
-	// reboot
-	taskid, err := d.driver.NodesNodeQemuVMIDStatusRebootPost(d.Node, d.VMID)
-	if err != nil {
-		return err
-	}
-
-	err = d.driver.WaitForTaskToComplete(d.Node, taskid)
-
-	return err
+	return d.OperateVM("restart")
 }
 
 // Kill the VM immediately
 func (d *Driver) Kill() error {
-	if len(d.VMID) < 1 {
-		return errors.New("invalid VMID")
-	}
-
-	err := d.connectAPI()
-	if err != nil {
-		return err
-	}
-
-	// sanity check the UUID
-	config, err := d.driver.GetConfig(d.Node, d.VMID)
-	if err != nil {
-		return err
-	}
-
-	cVMMUUID := getUUIDFromSmbios1(config.Data.Smbios1)
-	if len(d.VMUUID) > 1 && d.VMUUID != cVMMUUID {
-		return fmt.Errorf("UUID mismatch - %s (stored) vs %s (current)", d.VMUUID, cVMMUUID)
-	}
-
-	// stop
-	taskid, err := d.driver.NodesNodeQemuVMIDStatusStopPost(d.Node, d.VMID)
-	if err != nil {
-		return err
-	}
-
-	err = d.driver.WaitForTaskToComplete(d.Node, taskid)
-
-	return err
+	return d.OperateVM("kill")
 }
 
 // Remove removes the VM
 func (d *Driver) Remove() error {
-	if len(d.VMID) < 1 {
-		return nil
-	}
-
-	err := d.connectAPI()
+	vm, err := d.GetVM()
 	if err != nil {
 		return err
 	}
 
-	// sanity check the UUID
-	config, err := d.driver.GetConfig(d.Node, d.VMID)
-	if err != nil {
-		return err
+	stopTask, err2 := vm.Stop(context.Background())
+	if err2 != nil {
+		return err2
+	}
+	// wait for the stop task
+	if err3 := stopTask.Wait(context.Background(), time.Duration(5*time.Second), time.Duration(300*time.Second)); err3 != nil {
+		return err3
 	}
 
-	cVMMUUID := getUUIDFromSmbios1(config.Data.Smbios1)
-	if len(d.VMUUID) > 1 && d.VMUUID != cVMMUUID {
-		return nil
+	deleteTask, err4 := vm.Delete(context.Background())
+	if err4 != nil {
+		return err4
 	}
 
-	// force shut down VM before invoking delete
-	err = d.Kill()
-	if err != nil {
-		return err
+	// wait for the delete task
+	if err5 := deleteTask.Wait(context.Background(), time.Duration(5*time.Second), time.Duration(300*time.Second)); err5 != nil {
+		return err5
 	}
 
-	taskid, err := d.driver.NodesNodeQemuVMIDDelete(d.Node, d.VMID)
-	if err != nil {
-		return err
-	}
-
-	err = d.driver.WaitForTaskToComplete(d.Node, taskid)
-	return err
-}
-
-// Upgrade is currently a NOOP
-func (d *Driver) Upgrade() error {
 	return nil
 }
 
-// NewDriver returns a new driver
-func NewDriver(hostName, storePath string) drivers.Driver {
-	return &Driver{
-		BaseDriver: &drivers.BaseDriver{
-			SSHUser:     "docker",
-			MachineName: hostName,
-			StorePath:   storePath,
-		},
+func (d *Driver) GetVmidInRange() (int, error) {
+	// split d.VMIDRange into two parts by separating through ":"
+	vmidRange := strings.Split(d.VMIDRange, ":")
+	if len(vmidRange) != 2 {
+		return 0, fmt.Errorf("VMIDRange must be in the form of <min>:<max>. Given: %s", d.VMIDRange)
 	}
+
+	min, err := strconv.Atoi(vmidRange[0])
+	if err != nil {
+		return 0, err
+	}
+
+	max, err := strconv.Atoi(vmidRange[1])
+	if err != nil {
+		return 0, err
+	}
+
+	if min > max {
+		return 0, fmt.Errorf("VMIDRange :<max> must be greater than <min>. Given: %s", d.VMIDRange)
+	}
+	// now randomly choose from min max range
+	return rand.Intn(max-min) + min, nil
 }
 
 func (d *Driver) createSSHKey() (string, error) {
-	sshKeyPath := d.ResolveStorePath("id_rsa")
-	if err := mssh.GenerateSSHKey(sshKeyPath); err != nil {
+	var sshKeyPath = d.GetSSHKeyPath()
+	d.debugf("Creating SSH key at %s", sshKeyPath)
+
+	if err := ssh.GenerateSSHKey(sshKeyPath); err != nil {
 		return "", err
 	}
-	key, err := ioutil.ReadFile(sshKeyPath + ".pub")
+
+	key, err := os.ReadFile(sshKeyPath + ".pub")
+	d.debugf("Read SSH key from %s: %s", sshKeyPath, key)
 	if err != nil {
 		return "", err
 	}
 	return string(key), nil
-}
-
-// GetKeyPair returns a public/private key pair and an optional error
-func GetKeyPair(file string) (string, string, error) {
-	// read keys from file
-	_, err := os.Stat(file)
-	if err == nil {
-		priv, err := ioutil.ReadFile(file)
-		if err != nil {
-			fmt.Printf("Failed to read file - %s", err)
-			goto genKeys
-		}
-		pub, err := ioutil.ReadFile(file + ".pub")
-		if err != nil {
-			fmt.Printf("Failed to read pub file - %s", err)
-			goto genKeys
-		}
-		return string(pub), string(priv), nil
-	}
-
-	// generate keys and save to file
-genKeys:
-	pub, priv, err := GenKeyPair()
-	err = ioutil.WriteFile(file, []byte(priv), 0600)
-	if err != nil {
-		return "", "", fmt.Errorf("Failed to write file - %s", err)
-	}
-	err = ioutil.WriteFile(file+".pub", []byte(pub), 0644)
-	if err != nil {
-		return "", "", fmt.Errorf("Failed to write pub file - %s", err)
-	}
-
-	return pub, priv, nil
-}
-
-// GenKeyPair returns a freshly created public/private key pair and an optional error
-func GenKeyPair() (string, string, error) {
-	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		return "", "", err
-	}
-
-	privateKeyPEM := &pem.Block{Type: "RSA PRIVATE KEY", Bytes: x509.MarshalPKCS1PrivateKey(privateKey)}
-	var private bytes.Buffer
-	if err := pem.Encode(&private, privateKeyPEM); err != nil {
-		return "", "", err
-	}
-
-	// generate public key
-	pub, err := ssh.NewPublicKey(&privateKey.PublicKey)
-	if err != nil {
-		return "", "", err
-	}
-
-	public := ssh.MarshalAuthorizedKey(pub)
-	return string(public), private.String(), nil
-}
-
-func getUUIDFromSmbios1(str string) string {
-	var re = regexp.MustCompile(`(?m)uuid=([\d\w-]{1,})[,]{0,1}.*$`)
-	return re.FindStringSubmatch(fmt.Sprintf("%s", str))[1]
 }
